refactoring image savers
